fixed some things
